Kludge to address #12 (first row in Cursor null after upgrade)
Not really a fix. Works because the first call to importAssets() comes from the
PocketbookContent constructor, with overWrite == false, so the upgrade is not
done. Once onUpgrade() is called (presumably from getReadableDatabase()),
overWrite == true, the existing DB is deleted, and the new DB loaded. But
testing reveals that File.delete() doesn't happen until the process shuts down.
Regardless, onUpgrade() is only called once, and the File.delete() call allows
the new DB to be copied from assets/ without overwrite problems.